+++ database/migrations/2023_09_23_111809_remove_destination_from_services_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
@@ -12,8 +13,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('services', function (Blueprint $table) {
-            $table->dropColumn('destination_type');
-            $table->dropColumn('destination_id');
+            if (DB::connection()->getDriverName() !== 'sqlite') {
+                $table->dropColumn('destination_type');
+                $table->dropColumn('destination_id');
+            }
         });
     }
